add: db users with what was left in account and deleted entire accont service
