--- services/account/app/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "account",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "dev": "nodemon -q",
-    "start": "tsc && node dist/server.js",
-    "test:setup": "rm -f /usr/data/testDB.db && /usr/local/bin/testDB.sh",
-    "test": "npm run test:setup && DATABASE_PATH=/usr/data/testDB.db jest"
-  },
-  "author": "",
-  "license": "ISC",
-  "type": "module",
-  "dependencies": {
-    "@fastify/cors": "^11.1.0",
-    "@types/node": "^24.3.1",
-    "bcrypt": "^6.0.0",
-    "fastify": "^5.5.0",
-    "sqlite": "^5.1.1",
-    "sqlite3": "^5.1.7",
-    "typescript": "^5.9.3"
-  },
-  "devDependencies": {
-    "@types/bcrypt": "^6.0.0",
-    "@types/jest": "^30.0.0",
-    "jest": "^30.2.0",
-    "nodemon": "^3.1.10",
-    "pino-pretty": "^13.1.2",
-    "ts-jest": "^29.4.5"
-  }
-}
