--- services/account/app/src/app.ts
@@ -1,20 +1,0 @@
-import fastify, { type FastifyInstance } from 'fastify';
-import cors from '@fastify/cors';
-import { accountRoutes } from './routes.js';
-
-import { options } from './serv.conf.js';
-
-export async function buildServer(): Promise<FastifyInstance> {
-    const serv: FastifyInstance = fastify(options);
-
-    serv.register(cors, {
-        origin: true,
-        methods: ['GET', 'POST', 'PUT', 'DELETE'],
-    });
-
-    serv.register(dbConnector);
-    serv.register(accountRoutes);
-
-    await serv.ready();
-    return serv;
-}
