--- services/account/app/src/serv.d.ts
@@ -1,15 +1,0 @@
-import { Database } from 'sqlite';
-import 'fastify';
-
-declare module 'fastify' {
-	export interface FastifyInstance {
-		dbAccount: Database;
-	}
-
-	export interface FastifyRequest {
-		user: {
-			userID: number;
-			username: string;
-		};
-	}
-}
