--- services/account/tools/account.sh
@@ -1,8 +1,0 @@
-#!/bin/sh
-
-> /usr/app/server.log
-
-npm install -s
-
-npm run dev
-#npm start
